+++ src/backend/demo/management/commands/create_demo.py
@@ -27,15 +27,19 @@ def create_demo(stdout):
     )
 
     # Credentials
-    tx_models.ServiceProviderCredentials.objects.get_or_create(
+    source_service_credentials, _ = tx_models.ServiceProviderCredentials.objects.get_or_create(
         service_provider=source_service,
         client_id=os.environ.get("TOKEN_EXCHANGE_DEMO_SOURCE_CLIENT_ID"),
-        client_secret=os.environ.get("TOKEN_EXCHANGE_DEMO_SOURCE_CLIENT_SECRET"),
     )
-    tx_models.ServiceProviderCredentials.objects.get_or_create(
+    source_service_credentials.set_client_secret(
+        os.environ.get("TOKEN_EXCHANGE_DEMO_SOURCE_CLIENT_SECRET", "")
+    )
+    target_service_credentials, _ = tx_models.ServiceProviderCredentials.objects.get_or_create(
         service_provider=target_service,
         client_id=os.environ.get("TOKEN_EXCHANGE_DEMO_TARGET_CLIENT_ID"),
-        client_secret=os.environ.get("TOKEN_EXCHANGE_DEMO_TARGET_CLIENT_SECRET"),
+    )
+    target_service_credentials.set_client_secret(
+        os.environ.get("TOKEN_EXCHANGE_DEMO_TARGET_CLIENT_SECRET", "")
     )
 
     # Rules